requirements.txt
discord.py
pillow
aiohttp
qrcode